Migrate to using react router v6
