+++ login-workflow/example/.eslintrc.js
@@ -1,4 +1,5 @@
 module.exports = {
+    root: true,
     parser: '@typescript-eslint/parser',
     extends: ['@brightlayer-ui/eslint-config/tsx'],
     parserOptions: {
