+++ login-workflow/example/package.json
@@ -3,15 +3,15 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
+        "@brightlayer-ui/colors": "^3.0.1",
+        "@brightlayer-ui/icons-mui": "^3.0.0-beta.0",
+        "@brightlayer-ui/react-components": "^6.0.0-alpha.3",
+        "@brightlayer-ui/react-themes": "^7.0.0-alpha.2",
         "@emotion/react": "^11.7.1",
         "@emotion/styled": "^11.6.0",
         "@mui/icons-material": "^5.3.1",
         "@mui/material": "^5.4.1",
         "@mui/styles": "^5.4.1",
-        "@brightlayer-ui/colors": "^3.0.1",
-        "@brightlayer-ui/icons-mui": "^3.0.0-beta.0",
-        "@brightlayer-ui/react-components": "^6.0.0-alpha.3",
-        "@brightlayer-ui/react-themes": "^7.0.0-alpha.2",
         "@testing-library/jest-dom": "^5.0.0",
         "@testing-library/react": "^12.1.4",
         "@testing-library/user-event": "^14.0.0",
@@ -25,7 +25,7 @@
         "react-app-polyfill": "^3.0.0",
         "react-dom": "^17.0.0",
         "react-i18next": "^11.7.0",
-        "react-router-dom": "^5.2.0",
+        "react-router-dom": "^6.3.0",
         "react-scripts": "^5.0.0",
         "typescript": "^4.3.2"
     },
@@ -65,8 +65,8 @@
         "@typescript-eslint/parser": "^5.17.0",
         "enzyme": "^3.11.0",
         "eslint": "^8.12.0",
-        "eslint-config-prettier": "^8.3.0",
-        "eslint-plugin-react": "^7.24.0",
+        "eslint-config-prettier": "^8.5.0",
+        "eslint-plugin-react": "^7.29.4",
         "jest": "27.5.1",
         "prettier": "^2.3.1",
         "yarn-audit-fix": "^9.2.1"
