+++ login-workflow/example/src/App.tsx
@@ -23,6 +23,8 @@ import productLogo from './assets/images/eaton_stacked_logo.png';
 import i18n from 'i18next';
 import Button from '@mui/material/Button';
 import { useTranslation } from 'react-i18next';
+import { Navigate, Route } from 'react-router-dom';
+import { ExamplePreAuth } from './screens/ExamplePreAuth';
 
 // Uncomment these lines to add your app-specific translation resource
 i18n.addResourceBundle('en', 'app', { BUTTONLABEL: 'Change Language' });
@@ -124,9 +126,13 @@ export const App: React.FC = () => (
         <AuthUIConfiguration>
             <AuthNavigationContainer
                 routeConfig={routes}
-                // extraRoutes={[<Route exact path={'/pre-auth'} component={ExamplePreAuth} key={'pre-auth'} />]}
+                extraRoutes={[<Route path={'/pre-auth'} element={ExamplePreAuth} key={'pre-auth'} />]}
             >
-                <ExampleHome />
+                <Route path={''} element={<ExampleHome />} />
+                <Route path={'subpage'} element={<ExamplePreAuth />} />
+                <Route path={'subpage/nested'} element={<h1>Nested</h1>} />
+                {/* Redirect any other routes to home page */}
+                <Route path={'*'} element={<Navigate to={''} />} />
             </AuthNavigationContainer>
         </AuthUIConfiguration>
     </SecurityContextProvider>
