+++ login-workflow/package.json
@@ -42,7 +42,7 @@
         "react": "^16.13.1 || ^17.0.0",
         "react-dom": "^16.8.6 || ^17.0.0",
         "react-i18next": "^11.3.4",
-        "react-router-dom": "^5.2.0"
+        "react-router-dom": "^6.0.0"
     },
     "repository": {
         "type": "git",
@@ -96,8 +96,8 @@
         "enzyme-to-json": "^3.6.2",
         "eslint": "^8.12.0",
         "eslint-config-prettier": "^8.3.0",
-        "eslint-plugin-react": "^7.24.0",
-        "eslint-plugin-react-hooks": "^4.2.0",
+        "eslint-plugin-react": "^7.29.4",
+        "eslint-plugin-react-hooks": "^4.4.0",
         "i18next": "^20.3.1",
         "identity-obj-proxy": "^3.0.0",
         "jest": "27.5.1",
@@ -106,7 +106,7 @@
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "react-i18next": "^11.10.0",
-        "react-router-dom": "^5.2.0",
+        "react-router-dom": "^6.0.0",
         "regenerator-runtime": "^0.13.7",
         "ts-jest": "^27.1.4",
         "typescript": "^4.3.2"
