+++ login-workflow/src/components/Components.test.tsx
@@ -6,8 +6,7 @@ import Adapter from 'enzyme-adapter-react-16';
 import { BrandedCardContainer } from './BrandedCardContainer';
 import { FinishState } from './FinishState';
 import CheckCircle from '@mui/icons-material/CheckCircle';
-import { PrivateRoute } from './PrivateRoute';
-import { BrowserRouter } from 'react-router-dom';
+import { AuthGuard } from './AuthGuard';
 import { SecureTextField } from './SecureTextField';
 import { SimpleDialog } from './SimpleDialog';
 import { Spinner } from './Spinner';
@@ -83,23 +82,24 @@ describe('FinishState tests', () => {
 });
 
 jest.mock('@brightlayer-ui/react-auth-shared', () => ({
+    // @ts-ignore
     ...jest.requireActual('@brightlayer-ui/react-auth-shared'),
     useSecurityState: jest.fn().mockReturnValue({ isAuthenticatedUser: false }),
 }));
 
 jest.mock('react-router-dom', () => ({
+    // @ts-ignore
     ...jest.requireActual('react-router-dom'),
+    useNavigate: jest.fn().mockReturnValue(jest.fn()),
     useLocation: jest.fn().mockReturnValue('test-location'),
 }));
 
-describe('PrivateRoute unauthenticated tests', () => {
+describe('AuthGuard unauthenticated tests', () => {
     it('renders without crashing', () => {
         const div = document.createElement('div');
         ReactDOM.render(
             <ThemeProvider theme={theme}>
-                <BrowserRouter>
-                    <PrivateRoute authRoute={null} />
-                </BrowserRouter>
+                <AuthGuard authRoute={'/login'}></AuthGuard>
             </ThemeProvider>,
             div
         );
@@ -108,18 +108,17 @@ describe('PrivateRoute unauthenticated tests', () => {
 });
 
 jest.mock('@brightlayer-ui/react-auth-shared', () => ({
+    // @ts-ignore
     ...jest.requireActual('@brightlayer-ui/react-auth-shared'),
     useSecurityState: jest.fn().mockReturnValue({ isAuthenticatedUser: true }),
 }));
 
-describe('PrivateRoute authenticated tests', () => {
+describe('AuthGuard authenticated tests', () => {
     it('renders without crashing', () => {
         const div = document.createElement('div');
         ReactDOM.render(
             <ThemeProvider theme={theme}>
-                <BrowserRouter>
-                    <PrivateRoute authRoute={null} />
-                </BrowserRouter>
+                <AuthGuard authRoute={'/login'}></AuthGuard>
             </ThemeProvider>,
             div
         );
@@ -190,6 +189,7 @@ describe('ChangePasswordForm tests', () => {
 });
 
 jest.mock('@brightlayer-ui/react-auth-shared', () => ({
+    // @ts-ignore
     ...jest.requireActual('@brightlayer-ui/react-auth-shared'),
     useSecurityState: jest.fn().mockReturnValue({ isShowingChangePassword: true }),
     initialTransitState: jest.fn().mockReturnValue({ transitSuccess: true }),
@@ -227,6 +227,7 @@ describe('ChangePasswordModal transitSuccess=true tests', () => {
 });
 
 jest.mock('@brightlayer-ui/react-auth-shared', () => ({
+    // @ts-ignore
     ...jest.requireActual('@brightlayer-ui/react-auth-shared'),
     useSecurityState: jest.fn().mockReturnValue({ isShowingChangePassword: true }),
     initialTransitState: jest.fn().mockReturnValue({ transitSuccess: false }),
