+++ login-workflow/src/index.ts
@@ -3,8 +3,8 @@
  All rights reserved.
  This code is licensed under the BSD-3 license found in the LICENSE file in the root directory of this source tree and at https://opensource.org/licenses/BSD-3-Clause.
  **/
-export { AuthNavigationContainer } from './screens/AuthNavigationContainer';
-export { RouteConfig } from './screens/AuthNavigationContainer';
+export { AuthNavigationContainer } from './routing/AuthNavigationContainer';
+export { RouteConfig } from './routing/AuthNavigationContainer';
 
 export { default as i18n } from './translations/i18n';
